+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0 (June 11, 2025)
+
+* Added `Delete Object By ID` Action
+* Added `Lookup Object By ID` Action
+
 # 1.0.0 (June 03, 2025)
 
 * Initial component release